requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
pydantic
python-multipart
pandas
pyyaml
passlib
